Don't use default macos/arm64 deployment target in calculating the platform tag for fat binaries (pypa/wheel#390)
The system compiler in Xcode 12 will not set the deployment target for arm64 below 11.0.0
(which is the first version of macOS supporting arm64). To allow building wheels that target
an earlier version of macOS (by way of the x86_64 part of fat binaries) ignore the deployment
target in the arm64 part of fat binaries when that's 11.0.0.