Use the distribution object to run commands in build_meta
Once we have access to the distribution object, it is no longer
necessary to fully run the setup script.
Instead the backend can have more control over the process.

This change build on top of the `_get_dist` function and the
`distutils.core.run_commands` to trigger commands on the distribution
object.

For the legacy backend the setup script still runs traditionally
(but the corner cases are handled directly by `distutils.core.run_setup`)