Make `script_name` on `build_meta` less error prone
`distutils.sdist:sdist._add_defaults_standards` will try to add `script_name`
to the list of files to include in the distribution.

If the SETUP_SCRIPT does not exist, the best is to use some made-up name
that is unlikely to exist as a file.

If we use `sys.argv[0]` setuptools might try to add things like
`.tox/python/bin/pytest` to the tar.gz (and this will result in an error)