Move {pkg_resources,setuptools}/tests to the tests folder
This make it easy to deal with running the tests by installing the
distribution file.

(Originally a problem in PyPy was identified when trying to do that.
It is something tricky: e.g. the `setuptools` package is comming from
`site-packages` and `setuptools.tests` is comming from the current
working dir; moreover, `setuptools` is not a namespace package that
allow separation subpackages in multiple directories...)