Use in-memory encode detecting mechanism for distutils.core.exec patch
In `build_meta` we are patching `distutils.core` to make sure the latest
improvements in `distutils` are used even when it is loaded from stdlib.

One of these improvements is to me able to automatically detect encoding
for `setup.py` files via stdlib's `tokenize`. Previously, this process
was using a temporary file. The changes implemented here instead use
in-memory buffers (effectively replicating part of what `tokenize.open`
implements).

Co-authored-by: Sviatoslav Sydorenko <wk.cvs.github@sydorenko.org.ua>